Upload file added for reply/forward message

+++ app/controllers/conversations_controller.rb
@@ -1,11 +1,20 @@
 class ConversationsController < ApplicationController
   
   before_action :authenticate_user!
+  before_action :set_staff_list, only: [:show, :new, :edit_draft]
   
   def new
+    @subj=nil
+    @bod=nil
+    @recv=nil
   end
   
   def edit_draft
+    @subj= Mailboxer::Conversation.where(id: params[:id]).first.subject  
+    @bod= Mailboxer::Notification.where(id: Mailboxer::Conversation.where(id: params[:id]).first.receipts.first.notification_id).first.body
+    @recv= Mailboxer::Conversation.where(id: params[:id]).first.receipts.where(mailbox_type: 'inbox').pluck(:receiver_id)
+    notify_id=Mailboxer::Conversation.where(id: params[:id]).first.receipts.first.notification_id
+    @uploaded_files=AttachmentUploader.where(msgnotification_id: notify_id)
   end
   
   def create
@@ -21,21 +30,21 @@ class ConversationsController < ApplicationController
           newattach.msgnotification_id=notify_id  
           newattach.save
           if params[:submit_button]
-            flash[:notice]="mesej sent c/w attachment"
+            flash[:notice]=(t 'conversation.message_sent')
             redirect_to conversation_path(conversation)
           #########
           else
-            flash[:notice]="Selected file uploaded, add another file or click 'Send Message' to submit message!"
+            flash[:notice]=t 'conversation.file_uploaded_send_message'
             redirect_to action: 'edit_draft', id: conversation.id
           end  
           #########
         else
           notify_todraft=Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: true)
-          flash[:notice]="Attach file is invalid"
+          flash[:notice]=(t 'conversation.uploaded_invalid')
           redirect_to action: 'edit_draft', id: conversation.id
         end
       else
-        flash[:success] =(t 'conversation.message_sent')
+        flash[:success]=(t 'conversation.message_sent')
         redirect_to conversation_path(conversation)
       end   
     else
@@ -56,41 +65,88 @@ class ConversationsController < ApplicationController
          newattach.save
          if params[:submit_button]
            Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: false) #send draft (with attachment)
-           flash[:notice]="mesej sent c/w attachment"+"yea yea"
+           flash[:notice]=(t 'conversation.message_sent')
            redirect_to conversation_path(conversation)
          #########
          else
-           flash[:notice]="Selected file uploaded, add another file or click 'Send Message' to submit message! yea yea"
+           flash[:notice]=t 'conversation.file_uploaded_send_message'
            redirect_to action: 'edit_draft', id: conversation.id
          end  
          #########
        else
-         flash[:notice]="Attach file is Invalid yea yea"
+         
+         flash[:notice]=(t 'conversation.uploaded_invalid')
          redirect_to action: 'edit_draft', id: conversation.id
        end
      else #no attachment
        Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: false) 
-       flash[:success] =(t 'conversation.message_sent')+"yea yea"
+       flash[:success] =(t 'conversation.message_sent')
        redirect_to conversation_path(conversation)
      end
    end
 
   def show
-    @receipts = conversation.receipts_for(current_user)
+    drafts_notify_ids=Mailboxer::Notification.where(draft: true).pluck(:id)
+    @receipts_all = conversation.receipts_for(current_user)
+    @receipts=@receipts_all.where('notification_id NOT IN(?)', drafts_notify_ids)
     # mark conversation as read
     conversation.mark_as_read(current_user)
+    #default forward text
+    fw=Mailboxer::Notification.where(conversation_id: conversation.id).last
+    forward_body=fw.body
+    recipient=fw.receipts.where(mailbox_type: 'inbox').first.receiver.userable.name
+    details=" ("+(t 'from')+":"+fw.sender.userable.name+", "+(t 'to2')+":"+recipient+", "+(t 'conversation.subject')+":"+fw.subject+", "+(t 'conversation.date')+":"+fw.created_at.strftime("%A, %b %d, %Y at %I:%M%p")+")   "+(t 'conversation.message').upcase+": "
+    @forward_text=(t 'conversation.forwarded_message')+details+forward_body +" ---------------------------------------"
   end
   
   def reply
+    #raise params.inspect
     method=params["reply"]
     receipt_id=params["receiptid"]
     receipt=Mailboxer::Receipt.where(id: receipt_id).first
     if method=="1"
       if (!message_params[:body].blank?)
         #current_user.reply_to_conversation(conversation, message_params[:body])
-        current_user.reply_to_sender(receipt, message_params[:body])
-        flash[:notice]=t 'conversation.message_replied'
-        redirect_to conversation_path(conversation)
+        if message_params[:notify_id].nil? #new form
+          replied=current_user.reply_to_sender(receipt, message_params[:body]) 
+        #####
+          notify_id=replied.conversation.receipts.last.notification_id ####
+        else #draft form
+          notify_id=message_params[:notify_id].to_i
+        end
+        #########################################
+        unless message_params[:data].nil? #when attachment exist
+          newattach = AttachmentUploader.new
+          newattach.data = message_params[:data]
+          if newattach.valid?
+            newattach.msgnotification_id=notify_id  
+            newattach.save
+            if params[:submit_button]
+              Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: false) if message_params[:notify_id] #send draft (with attachment)
+              flash[:notice]=(t 'conversation.message_replied')
+              redirect_to conversation_path(conversation)
+            #########
+            else
+	      notify_todraft=Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: true)
+              flash[:notice]=(t 'conversation.file_uploaded_send_message')
+              redirect_to action: 'show', id: conversation.id  ###
+            end  
+            #########
+          else
+	    notify_todraft=Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: true) 
+            flash[:notice]=(t 'conversation.uploaded_invalid')
+            redirect_to action: 'show', id: conversation.id
+          end
+        else #no attachment
+          Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: false) if message_params[:notify_id] 
+          flash[:success] =(t 'conversation.message_replied')
+          redirect_to conversation_path(conversation)
+        end
+	############################################
+	
+	#####
+        #flash[:notice]=t 'conversation.message_replied'
+        #redirect_to conversation_path(conversation)
       else
         flash[:notice]=t 'conversation.message_required'
         redirect_to conversation_path(conversation)
@@ -100,17 +156,57 @@ class ConversationsController < ApplicationController
       recipients = User.where(id: message_params[:recipients]-[" "])
       if message_params[:recipients].count > 1 && !message_params[:body].blank?
 
-        # TODO - upon removal of conversation record, set pointer in DB (table Mailboxer::Conversation) to last record - ref: duplicates keys / violates related note
-        # NOTE Forwarding Message - temp solution 27Feb2016 
-        #1)send as new message first, 
-        #2)then update conversation ID in Notification same as current conversation
-        #3)remove newly created conversation
-        conversation2 = current_user.send_message(recipients, message_params[:body], conversation.subject).conversation
-        Mailboxer::Notification.where(conversation_id: conversation2.id).first.update_attributes(conversation_id: conversation.id)
-        Mailboxer::Conversation.where(id: conversation2.id).first.destroy
+        if message_params[:notify_id].nil? #new form
+          # TODO - upon removal of conversation record, set pointer in DB (table Mailboxer::Conversation) to last record - ref: duplicates keys / violates related note
+          # NOTE Forwarding Message - temp solution 27Feb2016 
+          #1)send as new message first, 
+          #2)then update conversation ID in Notification same as current conversation
+          #3)remove newly created conversation
+          #conversation2 = current_user.send_message(recipients, message_params[:body], conversation.subject).conversation
+          #Mailboxer::Notification.where(conversation_id: conversation2.id).first.update_attributes(conversation_id: conversation.id)
+          #Mailboxer::Conversation.where(id: conversation2.id).first.destroy
+          
+          conversation2 = current_user.send_message(recipients, message_params[:body], conversation.subject).conversation
+          Mailboxer::Notification.where(conversation_id: conversation2.id).first.update_attributes(conversation_id: conversation.id)
+          Mailboxer::Conversation.where(id: conversation2.id).first.destroy
+          #notify_id=Mailboxer::Notification.where(conversation_id: conversation2.id).first.id
+          #replied.conversation.receipts.last.notification_id
+          notify_id=conversation.receipts.last.notification_id # NOTE above updated record, use conversation instead of conversation2
+        else #draft form
+          notify_id=message_params[:notify_id].to_i
+        end
+        #########################################
+        unless message_params[:data].nil? #when attachment exist
+          newattach = AttachmentUploader.new
+          newattach.data = message_params[:data]
+          if newattach.valid?
+            newattach.msgnotification_id=notify_id  
+            newattach.save
+            if params[:submit_button]
+              Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: false) if message_params[:notify_id] #send draft (with attachment)
+              flash[:notice]=(t 'conversation.message_replied')+"huhu"
+              redirect_to conversation_path(conversation)
+            #########
+            else
+	      Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: true)
+              flash[:notice]=(t 'conversation.file_uploaded_send_message')+"hihi"
+              redirect_to action: 'show', id: conversation.id  ###
+            end  
+            #########
+          else
+	    notify_todraft=Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: true) 
+            flash[:notice]=(t 'conversation.uploaded_invalid')+"haha"
+            redirect_to action: 'show', id: conversation.id
+          end
+        else #no attachment
+          Mailboxer::Notification.where(id: notify_id).first.update_attributes(draft: false) if message_params[:notify_id] 
+          flash[:success] =(t 'conversation.message_replied')+"hoho"
+          redirect_to conversation_path(conversation)
+        end
+	############################################
 
-        flash[:notice] =(t 'conversation.message_forwarded')
-        redirect_to conversation_path(conversation)
+        #flash[:notice] =(t 'conversation.message_forwarded')
+        #redirect_to conversation_path(conversation)
       else
         flash[:error] =t 'conversation.select_recipient_message'
         redirect_to(:back)
@@ -146,13 +242,20 @@ class ConversationsController < ApplicationController
 #   end
   
   private
+  
+  def set_staff_list
+    user_ids=User.where(userable_type: 'Staff').where('userable_id is not null').order(userable_id: :asc).pluck(:id)
+    staff_list=[]
+    user_ids.each{|user_id|  staff_list << [Staff.joins(:users).where('users.id=?', user_id).first.name, user_id]}
+    @staff_list=staff_list.sort
+  end
 
   def conversation_params
     params.require(:conversation).permit(:id, :subject, :body, :data, recipients:[])
   end
   
   def message_params
-    params.require(:message).permit(:body, :subject, recipients:[])
+    params.require(:message).permit(:body, :subject, :data, :notify_id, recipients:[])
   end
